WIP: find "bare" reaction events in function calls
Search attempt functions for calls to egre:attempt/2 and record the
second argument as a reaction event:

e.g.

attempt(_, _, ActionEvent, _) ->
   egre_object:attempt(Target, {foo, Bar, baz}).

  Here the action event is "ActionEvent" and the reaction event
  is {foo, Bar, baz}.

I also fixed up a test fixture so that the module name was binary, which
is what the parse transform code spits out. It's arbitrary when we turn
all the data into binary (since we're going to end up writing this
information out into an "event chain" text file) but currently we get
the module name as a binary and the function name as an atom.

Todo:
- add succeed/1 function handling where the API call argument is a
three-tuple, as opposed to attempt/1 where the argument is a four-tuple.

- succeed/1 functions are more likely to trigger an event with
egre_object:attempt/1 whereas attempt/1 functions will likely add new
events to the #result{} record. We need to search the #result{} record

  - Sometimes the attempt/1 reaction events are temporarily stored in
  "Event" or "NewEvent" variables, so we'll need to search for those.
  We can potentially just use "Event"/"NewEvent" variable matches if
  they exist and not bother with the #result{} record.

- track down type inference data within event field function calls:
  e.g. NewEvent = {foo(), Bar, baz}
  foo() should be inlined as a case statement so we'll need to trace
  through each clause to build up a list of potential events with their
  type inference:
  e.g. NewEvent = {case X of
                       1 -> <<"foo">>;
                       2 -> 99
                   end, Bar, baz}

      This works out to two different events:

      {1, 2, baz}
      [{1, binary}, {2, <whatever Bar is, if we know>}],
      [{1, <<"CaseStatement">>}, {2, "Baz"}]

      {1, 2, baz}
      [{1, int}, {2, <whatever Bar is, if we know>}],
      [{1, <<"CaseStatement">>}, {2, "Baz"}]

      i.e. one event where the case expression could evaluate to
      the binary <<"foo">> and one event where the case expression
      could evaluate to the int 99.