Capture resend, broadcast and modified events
Based on the #result.result field and the #result.event field we can
have different ways to kick off new events from an attempt/1 API call.

This includes logic to keep track of variables and then expand variables
that are used for the #result.result and #result.event field values.
(I think this _maybe_ should be inlined along with the function clauses
in the parse transform module.)

4 new tests:
- resend an event as a raw tuple
- resend an event as a variable containing a tuple
- broadcast an event as a raw tuple
- return a modified event as a raw tuple

We still need tests for broadcasting and returning events where the
event is stored in a variable.

I'm not sure if I'll dig into events contained in a variable where parts
of the event are also in variables: e.g.

Target = <some process>,
Event = {do, foo, to, Target}