egre_object:attempt shouldn't be used
EGRE rules modules should be using the egre:attempt/2,3 API call.

Also account for the possible 3rd argument to egre:attempt/2,3
specifying whether or not (typically not, since true is the default) to
subscribe to the generated (reaction) event.

Also rename the 2nd arg in egre:attempt/2,3 from "Arguments" to "Event"
because "Arguments" makes zero sense.