Support more type inference for event chains
Add new type inference cases:
- is_pid/1
- equality with integer, char, float, string, nil, list, etc.

Capture type of original var when assigning one var to another:
e.g. B = 1, A = B.
We know that A will have the same value as B, and the same type.

Note that [] is nil, "abc" is string, and [a, 1] is list.

Added some tests.

I used the following src to compile with the ID transform to get the AST
so I could test out what the AST types were for different cases.

  Compile with: erlc -pa test test_parse.erl

 ░ -module(test_parse).
 ░ -compile({parse_transform, egre_protocol_id_transform}).
 ░ -export([attempt/1]).
 ░
 ░ attempt(_A) ->
 ░     Int = 1,
 ░     Float = 1.0,
 ░     EmptyList = [],
 ░     Char = $c,
 ░     String = "string",
 ░     Bin = <<"bin">>,
 ░     Atom = atom,
 ░     List = [a, 1],
 ░     List2 = List,
 ░     _List3 = List2,
 ░     [Int, Float, EmptyList, Char, String, Bin, Atom, List, List2].