Infer types from +/- ops, infer types recursively
So far, EGRE_MUD_1 only uses integers in events, so I'm assuming any
plus/minus operands are integers.

If a literal "A + B" is in an event, then the "variable" name will be
"(A + B)" and the type will be integer, e.g.

                 | {{do, 1},
 {do, A + B} --> |  [{1, <<"(A + B)">>}],
                 |  [{1, integer}]}

  ... so, even though the expression is A + B, we count it as a single
  value in the event, because that's what it will be once it's sent out
  to other objects; however, we still track that "the" variable (i.e.
  the expression) was originally an expression.

Also make sure that we recurse into statements looking for type
inference, instead of just stopping, for instance, at a variable
assignment.

So far, my plan is to work on the MUD rules and the protocol chain parse
transform gear in parallel. There are only so many combinations of type
inference, and slowly get more and more.