WIP: split cases into multiple code paths
If a case statement as a case statement for an expression, the whole
outer case statement should be duplicated for each code path in the
expression case clause, including guards.

Further, if a case statement has a function for an expression, the
function should be inlined as a case clause and the above paragraph
should apply.

Since cases and functions (which become cases) can occur within other
code blocks, we have to recurse through those code blocks, e.g. tuples,
match, begin-end. The parent code block has to be duplicated for every
code path combination of the children.

e.g. {fun1(X), fun2(Y)}

  If fun1/1 and fun2/1 both have 2 clauses each, then there's a total of
  four possible code paths, and we'll need four couples of the tuple,
  each with one of the possible code paths.

This is all required because functions and cases can match on data
types, so each clause could lead to a code path where variables have
different types. When we generate an event, we need to know what types
of the variables were that went into that event; the only way we can
know that is if we follow all the possible code paths that could have
led to that event.

Why do we need to know the event types? Because some events have the
same shape, but are handled differently by rules modules based on type.
For instance, changing names into pids: same shape event, different
types (i.e. binary -> pid)

This is very close to working, but the "in" AST and the "out" AST seem
to be sorting differently.

TODO: traverse through begin-end blocks.