WIP: Event pairs need to be matched up into chains
An event pair consists of an action event and reaction event: i.e. what
happened, and how a rules module is responding to what happened.

Rules modules will feed events to each other: action event A will be
picked up by rules module 1 and trigger reaction event B. Module 2 might
see event B as its action event, and react with event C. These two
action-reaction pairs now form a chain:
A -> B
  B -> C

The goal of egre_protocol_event_chains.erl is to take the pairs that
egre_protocol_event_pairs.erl writes out to a BERT file, and link them
together in chains.

Reaction events are matched up action events (e.g. the B in A -> B to
the B in B -> C) by "shape" and types. The shape is always "normalized"
so that variable names or expressions are turned into ordinal numbers,
e.g. {1, foo, 2, bar, 3}
This is so that the same event shapes can match even if different
variable names are used in different places.

The logic to chain the event pairs together is done, and now we need to
write the chains out to file.

TODO:
- serialize event pair chains
- write chains out to file
- figure out if this actually works!