Tuple elements not being handled
Tuples in AST look like this: {tuple, _Elements = [...]}.

I was taking any unmatched form, which are all expressed as tuples, and
turning them into lists and looping over them; however, I didn't have
code to dig into lists, e.g. within a {tuple, _} form.

This changes matches on {tuple, Elements} and folds over the elements.

The reason this is important is because there was an event being
generated containing a function call in the constructed tuple, and the
call wasn't being inlined, because the tuple elements weren't being
processed.