Some calls have no (re)actions
If an attempt doesn't have a reaction, then its only reason for
existance is to subscribe to, or fail an event. If the attempt call
subscribes to the event, then a succeed call will also respond to that
event. Most reactions come from succeed calls.

If an api function (attempt/1, succeed/1) matches the event to an
ignored variable, then that "action" event is returned as an empty list,
and we don't need to bother making a pair for it.