Event tuples duplicating previous variables
I was both passing in the existing event to fold over a tuple inside of
an event, and then also appending the resulting event onto the original
event:

e.g.
  NextElement = {c, d, e}
  EventSoFar = [a, b]
  EventAfterTuple = [a, b, {a, b, c, d, e}].

If we create a new accumulator with an empty event, we can wrap the
resulting sub-event in a tuple, and then append it to the already
processed event elements.

e.g.
  NextElement = {c, d, e}
  EventSoFar = [a, b]
  EventForTuple = []     <-- replace current event in accumlator
  EventAfterTuple = [{c, d, e}]
  CombinedEvent = [a, b, {c, d, e}]