Types don't have to match exactly
If an element in one event doesn't have a type, and the corresponding
element in a potential map does have a type, then the types are
considered a match. i.e. '_' is "any type"

e.g. consider these pairs of event tuples, whose shapes match, and
     the effect of different type lists.

This is a match:

  {1, say, 2}  ['_', bin]
  {1, say, 2}, ['_', '_']

This is a match:

  {1, say, 2}  ['_', '_']
  {1, say, 2}, ['_', '_']

This is a match:

  {1, say, 2}  ['_', bin]
  {1, say, 2}, ['_', bin]

This is not a match:

  {1, say, 2}  ['_', list]
  {1, say, 2}, ['_', bin]