DRY out compilation
I was compiling the in and out versions of the code with basically the
same code, so I abstracted it out.

Reminder: "out" code is code that is written in Erlang with the
intention of turning into the expected forms with the ID parse
transform. "in" code is the normal code that will be transformed by the
parse transformer and should produce forms that match the "out" forms.